requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
pymysql
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
sqlacodegen
python-dotenv
httpx
alembic

python-socketio>=5.16
simple-websocket
